Implemented <Enter> key behavior
This commit was made in FreeBSD. Yes, you heard that right, FreeBSD...
So, I've been trying out some OSes for a while now, such as TempleOS, Plan 9, and most recently FreeBSD.
I tried to code in TempleOS in HolyC and oh boy was it different than standard C, not to mention that there is not `git` in TempleOS (duh).
Then, came Plan 9. I dabbled a bit there and I can say that I'm not a fan of `rio`, the window management system of Plan 9.
`rio` was unintuitive for me. The fact that everything needs the use of a mouse disgusts me.
At last, came FreeBSD. Well, FreeBSD is also a Unix-like OS just like Linux (don't come to me and say "actually, what you're referring to as Linux, is in fact, GNU/Linux", I won't give a single fuck.) that uses the BSD license rather than the GPL license.
What surprised me is that the base package of FreeBSD doesn't contain `gcc`, it came with `clang` as the default compiler.