Implemented basic pattern search (CTRL-F behavior)
CTRL-F will now open up a prompt in which the user can type their search pattern.
When the supplied search pattern is found in the file, the cursor will jump to the first character of the search pattern and the cursor will jump to the top of the screen.
The user can now abort their search by pressing ESC.
There is also a new help status message shown when the program starts which says "Ctrl-F = find".