Implemented incremental search feature
CTRL-F behavior is now incremental.
In the past, the user needed to press <Enter> key for their search pattern to be handled and the result shown to them.
Now, every keypress in the search prompt is being handled and will start searching for the current search pattern provided to the prompt (No need to press <Enter> to see result).
Keep in mind that to quit the search prompt, the user will need to press <ESC> or <Enter>, else they will still be in the search prompt.